# Simplify deps