--- a/qi-lib/info.rkt
+++ b/qi-lib/info.rkt
@@ -3,10 +3,8 @@
 (define version "2.0")
 (define collection "qi")
 (define deps '("base"
-               "mischief"
                "fancy-app"
-               "typed-stack"
-               "adjutor"))
+               "typed-stack"))
 (define build-deps '())
 (define clean '("compiled" "private/compiled"))
 (define pkg-authors '(countvajhula))