--- a/qi-lib/info.rkt
+++ b/qi-lib/info.rkt
@@ -3,8 +3,7 @@
 (define version "2.0")
 (define collection "qi")
 (define deps '("base"
-               "fancy-app"
-               "typed-stack"))
+               "fancy-app"))
 (define build-deps '())
 (define clean '("compiled" "private/compiled"))
 (define pkg-authors '(countvajhula))