@countvajhula 2022-07-10
This is what I get:

```
BEFORE:
real	0m0.850s
real	0m0.807s
real	0m0.894s
real	0m0.830s
real	0m0.812s

AFTER:
real	0m0.804s
real	0m0.787s
real	0m0.795s
real	0m0.798s
real	0m0.802s
```

```
> (~> (0.850 0.807 0.894 0.830 0.812) (-< + count) /)
            
0.8386000000000001

> (~> (0.804 0.787 0.795 0.798 0.802) (-< + count) /)
            
0.7972

```

Looks like about a 40ms or about 5% improvement 😄