@benknoble 2022-07-10
Wow, I was expecting a bigger difference. I'm assuming you ran `raco setup`?