@countvajhula 2022-07-11
Yeah me too! But yes, I did run raco setup. I wonder what is causing the majority of the latency then, if not mischief. I’ll try running this test for some other libraries and see what times look like.

Sent from my iPhone

> On Jul 10, 2022, at 4:33 PM, D. Ben Knoble ***@***.***> wrote:
> 
> Wow, I was expecting a bigger difference. I'm assuming you ran raco setup?
> 
> —
> Reply to this email directly, view it on GitHub, or unsubscribe.
> You are receiving this because you were mentioned.