@countvajhula 2022-07-12
I wrote a [small CLI script](https://gist.github.com/countvajhula/33b43c488c9e384e87420140597783a4) that accepts a module name at the command line and then:
- creates a Racket file in `/tmp/` that has the structure that @sarna suggested, including a `(require module-name)` using the module name provided
- also creates a baseline file that has the same structure but without the `require`
- Runs `time racket ...` on each of these
- Returns the difference of the two times, as an estimate of the time taken for the module specifically (i.e. factoring out `racket/base`)

Run it as e.g.:
```
$ ./loadlib.rkt rebellion
```

The times on packages I tested ranged 0-2 s, with the majority in the range 0-500 ms, so Qi being on the higher end, as reported.

`mischief` came in heavy at 570 ms, while `mischief/shorthand` by itself was 150 ms, explaining why it wasn't contributing as much as we thought. Adjutor is about 200 ms, so you'd think it would trim that much off the `(require qi)` loading time, so, not sure why it only trimmed < 40 ms.

Interestingly enough, `rebellion` appears to have no overhead over `racket/base` at all, not sure how @notjack managed this magic!