@benknoble 2022-07-12
@countvajhula after the tests pass (🤞) could you try profiling this again? I expect this to make a bigger difference now that the deps for `qi-lib` are slimmed as much as possible.