@benknoble 2022-07-12
Heh, when installing the package takes 6+ minutes but running the tests takes 14s… (the CS current build for the "remove typed-stack dependency" commit https://github.com/countvajhula/qi/pull/51/commits/11d5ec8562a28ec1d18258aadf004e12db4b8ada)