@countvajhula 2022-07-12
So, yes, basically, a factor of 3 improvement. No big deal at all 😆 

Here's the full data from running the scripts.

1. First, the dependency breakdown using [`loadprof`](https://gist.github.com/Bogdanp/c36e9bf354917455dbb549be9ffb4906) from @Bogdanp :

**Before**:

```
$ ./loadprof.rkt qi/flow
...
Timings:
* /Users/siddhartha/work/lisp/racket/qi/qi-lib/flow.rkt (274.30ms)
  * ...Users/siddhartha/work/lisp/racket/qi/qi-lib/private/util.rkt (246.39ms)
    * (submod ...ddhartha/Library/Racket/8.3/pkgs/typed-stack/typed-stack.rkt #%contract-defs) (237.96ms)
    * ...sers/siddhartha/Library/Racket/8.3/pkgs/typed-stack/main.rkt (208.26ms)
    * /Users/siddhartha/Library/Racket/8.3/pkgs/adjutor/main.rkt (19.47ms)
      * ...a/Library/Racket/8.3/pkgs/adjutor/stable/require-provide.rkt (11.39ms)
      * ...dhartha/Library/Racket/8.3/pkgs/adjutor/stable/serialize.rkt (10.34ms)
      * ...ary/Racket/8.3/pkgs/adjutor/stable/environment-variables.rkt (7.02ms)
      * ...artha/Library/Racket/8.3/pkgs/adjutor/stable/define-star.rkt (6.15ms)
  * /Users/siddhartha/Library/Racket/8.3/pkgs/adjutor/main.rkt (19.47ms)
    * ...a/Library/Racket/8.3/pkgs/adjutor/stable/require-provide.rkt (11.39ms)
    * ...dhartha/Library/Racket/8.3/pkgs/adjutor/stable/serialize.rkt (10.34ms)
    * ...ary/Racket/8.3/pkgs/adjutor/stable/environment-variables.rkt (7.02ms)
    * ...artha/Library/Racket/8.3/pkgs/adjutor/stable/define-star.rkt (6.15ms)
  * /Users/siddhartha/work/lisp/racket/qi/qi-lib/macro.rkt (6.39ms)
```

**After**:

```
$ ./loadprof.rkt qi/flow
...
Timings:
* /Users/siddhartha/work/lisp/racket/qi/qi-lib/flow.rkt (11.31ms)
  * /Users/siddhartha/work/lisp/racket/qi/qi-lib/macro.rkt (5.83ms)
```

2. The dependency breakdown using the `-l` option, which shows the proportion of total time contributed by the leaves in the dependency tree, sorted by the ones taking up the most time:

**Before**:

```
Timings:
* (submod ...ddhartha/Library/Racket/8.3/pkgs/typed-stack/typed-stack.rkt #%contract-defs) (248.77ms, weight 0.341886)
* ...sers/siddhartha/Library/Racket/8.3/pkgs/typed-stack/main.rkt (221.54ms, weight 0.304462)
* ...Users/siddhartha/work/lisp/racket/qi/qi-lib/private/util.rkt (183.36ms, weight 0.251991)
* /Users/siddhartha/Library/Racket/8.3/pkgs/adjutor/main.rkt (19.92ms, weight 0.027375)
* ...dhartha/Library/Racket/8.3/pkgs/adjutor/stable/serialize.rkt (10.32ms, weight 0.014179)
* ...a/Library/Racket/8.3/pkgs/adjutor/stable/require-provide.rkt (8.13ms, weight 0.011178)
* /Users/siddhartha/work/lisp/racket/qi/qi-lib/macro.rkt (6.30ms, weight 0.008651)
* ...ary/Racket/8.3/pkgs/adjutor/stable/environment-variables.rkt (5.35ms, weight 0.007348)
```

**After**:
```
Timings:
* /Users/siddhartha/work/lisp/racket/qi/qi-lib/macro.rkt (6.46ms, weight 0.257713)
```

3. Using the [hyperfine](https://github.com/sharkdp/hyperfine) utility recommended by @samth , the naive cost of `racket -l qi` is:

**Before**:
```
$ ./hyperfine "racket -l qi"
Benchmark 1: racket -l qi
  Time (mean ± σ):     672.5 ms ±   5.9 ms    [User: 544.3 ms, System: 122.0 ms]
  Range (min … max):   664.5 ms … 682.3 ms    10 runs
```

**After**:
```
$ ./hyperfine "racket -l qi"
Benchmark 1: racket -l qi
  Time (mean ± σ):     347.1 ms ±   3.6 ms    [User: 261.7 ms, System: 80.0 ms]
  Range (min … max):   343.5 ms … 354.0 ms    10 runs
```

While this is statistically rigorous, it does not factor out the cost of `racket/base`, so it doesn't tell us the true extent of improvement from these changes. So:

4. Finally, the resulting output from [`loadlib`](https://gist.github.com/countvajhula/33b43c488c9e384e87420140597783a4) which is basically a descendent of @sarna 's method for measuring the cost of `(require qi)`:

**Before**:
`$ ./loadlib.rkt qi` (run five times):
```
500.0 ms
500.0 ms
500.0 ms
510.0 ms
520.0 ms
```

**After**:
`$ ./loadlib.rkt qi` (run five times):
```
170.0 ms
170.0 ms
160.0 ms
170.0 ms
180.00000000000003 ms
```

I'd say that's pretty darn convincing.