@countvajhula 2022-07-12
Yeah definitely. I'm wondering what the best way to collect these kinds of tools might be so that they're discoverable. Either a repo containing them all, or an "awesome X" style repo containing curated links, or ...?

In any case, this LGTM. Feel free to merge it if you think it's good to go :)