@benknoble 2022-07-12
I'm not planning on doing any more (_e.g._, installing the repo takes a while because of the deps for docs and profiling and such, but I'm not going to tackle that today and it may not be worth tackling), so since you gave the green-light I will indeed merge this.