@Bogdanp 2022-07-13
Just a heads-up that I've turned the gist into a proper package:

* https://github.com/Bogdanp/racket-import-profiler
* https://pkgd.racket-lang.org/pkgn/package/import-profiler