@countvajhula 2022-07-14
@sarna Any interest in maintaining something resembling `loadlib.rkt` as a package similar to Bogdan's `racket-import-profiler` above? I think it would be very similar in structure so a lot of the config could probably mirror `racket-import-profiler` almost verbatim. If not, I'm happy to maintain it as well, lmk.